% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{execute_py}
\alias{execute_py}
\title{Title run python file in R}
\usage{
execute_py(python_path = NULL, script_path = NULL, ...)
}
\arguments{
\item{python_path}{python interpreter}

\item{script_path}{python script}

\item{...}{sys.argv must be order}
}
\description{
Title run python file in R
}
