% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_GRN.R
\name{generate_GRN}
\alias{generate_GRN}
\title{Title utilize Qiuyue Yuan, Zhana Duren., \emph{Nat Biotechnol}(2024). Apr 12
and Wang et al., \emph{Nat Commun}(2023). Dec 20;14(1):8459
method to construct the TF-CRE-Target prior network}
\usage{
generate_GRN(
  peaks,
  Genes,
  input,
  out = "out",
  genome_ = "hg38",
  min_overlap = 400,
  Start_ali = F,
  Sparse = F,
  motifList = NULL,
  method = "FIMO",
  using_scan = "union",
  strings_width = 1000,
  ncores = 12,
  p_thresh = 1e-05,
  Gene_network = NULL
)
}
\arguments{
\item{peaks}{chr1-10000-11000}

\item{input}{the data from Qiuyue Yuan, Zhana Duren., \emph{Nat Biotechnol}(2024). Apr 12}

\item{genome_}{human & mouse}

\item{min_overlap}{the peak_a intersect with peak_b}

\item{motifList}{FIMO and motifmatchr}

\item{p_thresh}{motifScan 1e-4}

\item{Gene_network}{the data from Wang et al., \emph{Nat Commun}(2023). Dec 20;14(1):8459}
}
\description{
Title utilize Qiuyue Yuan, Zhana Duren., \emph{Nat Biotechnol}(2024). Apr 12
and Wang et al., \emph{Nat Commun}(2023). Dec 20;14(1):8459
method to construct the TF-CRE-Target prior network
}
\examples{
\dontrun{
generate_GRN(peaks = peaks,
Genes = Genes,
input = "/mnt/data/home/tycloud/workspace/paper_code/LINGER/data/data_bulk/",
min_overlap = 300,
Start_ali = F,
motifList = pfmList,
method = "FIMO",
ncores = 12,
p_thresh=1e-5,
using_scan=F,
Gene_network="/mnt/data/home/tycloud/workspace/algorithms_raw/data/network_human.csv",
network_out="/mnt/data/home/tycloud/workspace/algorithms_raw/data/multiome_prior_network.csv")
}
}
