% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_GRN.R
\name{generate_Motifpeaks_scan}
\alias{generate_Motifpeaks_scan}
\title{Title}
\usage{
generate_Motifpeaks_scan(
  peaks,
  meme_bin_path = "/mnt/data/home/tycloud/meme/bin",
  genome_,
  motifList = NULL,
  pfmList = NULL,
  method = "FIMO",
  strings_width = 1000,
  ncores = 12,
  p_thresh = 1e-05
)
}
\arguments{
\item{peaks}{peaks should be chr1-8888-8888}

\item{motifList}{if method=="FIMO", the length of motifList should
be < 1500}

\item{method}{Parallel computation of "motifmatchr"
has not been tested and is not recommended}
}
\description{
Title
}
