% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineageNet-package.R
\docType{package}
\name{LineageNet-package}
\alias{LineageNet}
\alias{LineageNet-package}
\title{LineageNet: Integrated spatio-temporal multi-omics modeling framework}
\description{
Integrated spatio-temporal multi-omics modeling framework.
}
\author{
\strong{Maintainer}: YuHong Huang \email{1395214203@qq.com} (PhD candidate)

Authors:
\itemize{
  \item Xiao Zhai \email{zhaix01@dmu.edu.cn} (Senior Engineer)
  \item Tao Song \email{songtao@dmu.edu.cn} (Director of Biomedical Engineering Faculty)
}

}
\keyword{internal}
