% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MAST_DEG}
\alias{MAST_DEG}
\title{Title Differential gene expression analysis by MAST}
\usage{
MAST_DEG(
  exp_data,
  pseudotime_lineages,
  split_num,
  log_ = T,
  out_dir = "result"
)
}
\arguments{
\item{exp_data}{col:cellid, row:gene/peak id}

\item{log_}{}

\item{pseudotime_}{should exports by slingshot}
}
\description{
Title Differential gene expression analysis by MAST
}
