% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_GRN.R
\name{peaks_overlap}
\alias{peaks_overlap}
\title{Title
peaks and peaks_ref should be chr1-8888-8888}
\usage{
peaks_overlap(peaks, peaks_ref, min_overlap, dt = T)
}
\arguments{
\item{peaks_ref}{}
}
\description{
Title
peaks and peaks_ref should be chr1-8888-8888
}
