% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_GRN.R
\name{peaks_to_string}
\alias{peaks_to_string}
\title{Title
peaks should be chr1-8888-8888}
\usage{
peaks_to_string(peaks, genome_)
}
\arguments{
\item{peaks}{}
}
\description{
Title
peaks should be chr1-8888-8888
}
