% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_trajectory_DAG}
\alias{plot_trajectory_DAG}
\title{Title plot_trajectory_DAG Implemented via the ggraph package}
\usage{
plot_trajectory_DAG(
  Tab_l,
  Tab_l2,
  scale_ = NULL,
  order_ = NULL,
  node_df = NULL,
  color_by = NULL,
  cellid_by = NULL,
  cut_ = 0.1,
  trans_link = 0.3,
  width_min = 0.1,
  width_max = 0.3,
  width_2 = 0.2,
  width_top = 0.5,
  alpha_min = 0.1,
  x_size = 3,
  y_size = 3,
  node_size = 2,
  node_color = "gray40",
  edge_color
)
}
\arguments{
\item{Tab_l}{the frist elemnet must be a single root,
The list element can't be all = 0,
the names of list must be splited by ":"}

\item{scale_}{"src", "obj" or NULL}

\item{order_}{the order in y axis: from low tp high}

\item{cut_}{The minimum transformation probability that satisfies the drawing}

\item{trans_link}{The value of links within same time; eg: E9_p,E9_l}

\item{node_color}{has been set in node_info}
}
\description{
Title plot_trajectory_DAG Implemented via the ggraph package
}
