% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotDot}
\alias{plotDot}
\title{use ggplot2 to plot heatmap and bubble, can cluster row or col, add row and col metadata}
\usage{
plotDot(
  object,
  features,
  model = 1,
  my_pal,
  my_pal_10 = c("#4292C6", "#EF3B2C"),
  my_pal_2 = c("#F7903D", "#4D85BD", "#59A95A"),
  my_mal_gradientn = NULL,
  layer = "data",
  ID = NULL,
  r_ID = NULL,
  out_ID = NULL,
  split_col = NULL,
  cluster_ID = TRUE,
  cluster_feature = TRUE,
  flip_ = F,
  plot_style = "bubble",
  white_line_size = 0.5,
  white_line_color = "white",
  legend_position = "rigth",
  legend_direction = "vertical",
  y_size = 10,
  x_size = 10,
  x_linewidth = 0.5,
  y_linewidth = 0.5,
  x_angle = 0,
  replace_ = F
)
}
\arguments{
\item{object}{a seurat object or data.frame,
if object is data.frame,it should be a two colmuns var(not a mat),
colinfo and rowinfo all should be in}

\item{model}{#' 1 metadata;
2 3 metadata;
3 Don't known how many data}

\item{ID}{'_' can't in ID}

\item{out_ID}{only used in model 1 or 2}

\item{cluster_feature}{must >2 for hcluster}

\item{replace_}{}
}
\description{
use ggplot2 to plot heatmap and bubble, can cluster row or col, add row and col metadata
}
\examples{
\dontrun{
plotDot(object = seob,
features=unique(c('Sox9', 'Col2a1',
                  'Pecam1','Cdh5',
                 'Krt14','Epcam',
                  'Plp1','Sox10',
                  'Mitf',
                  'Prrx1', 'Twist1',
                  'Myod1','Myog')),
mode=1, ID=c('celltype'),
out_ID = c('celltype'),
my_pal = colorRampPalette(c(my_pal_science6))(7),
my_mal_gradientn = c("#4C3399","#F5F2EF","#FFE714"),
cluster_ID = F,
cluster_feature = F,
x_angle=45)
}
}
