% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runmonocl3.R
\name{Runmonocl3}
\alias{Runmonocl3}
\title{Title a Warpper function of monocle3}
\usage{
Runmonocl3(
  seob,
  cell_metadata,
  useumap = "seurat",
  cluster = "seurat_clusters",
  use_ncenter = TRUE,
  ncenter = 300,
  num_dim = 50
)
}
\arguments{
\item{num_dim}{}
}
\description{
Title a Warpper function of monocle3
}
\examples{
\dontrun{
cds_E8_5_E9 = Runmonocl3(seob = seob_E8_5_E9,
cluster = "celltype_velocyto")
}
}
