% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSeurat.R
\name{RunSeurat}
\alias{RunSeurat}
\title{Title
A Wrapper of seurat integrated, reduce and cluster}
\usage{
RunSeurat(
  seob,
  split_by = NULL,
  coln_add = NULL,
  cellid_add = T,
  genes_add = NULL,
  integrated = NULL,
  integrated_assay = "RNA",
  cellid = NULL,
  min_cells = 0,
  min_features = 0,
  max_features = 0,
  MT_filter = 0,
  nfeatures = 3000,
  npcs = 30,
  resolution = 0.3,
  k_anchor = 5,
  k_filter = 200,
  k_score = 30,
  outfeatures = NULL,
  min_batch_num = NULL,
  k_weight = 100,
  use_old_RNA = F,
  use_old_Seurat = F,
  use_old_embeddings = F
)
}
\arguments{
\item{seob}{seurat object}

\item{k_filter}{Adjusting the value of k.filter can change
the number of cells considered as potential anchors.
Higher k.filter values imply stricter screening criteria,
which may reduce the number of anchors
but improve the specificity of integration;
lower k.filter values, on the other hand,
may increase the number of anchors,
which helps to capture a larger population of cells
but may increase false integrations}

\item{k_score}{The k.score parameter defines
the number of nearest neighbors to be used
when evaluating the quality of each anchor candidate.
This parameter is mainly used to evaluate the stability of
each anchor point:
For each potential anchor point, FindIntegrationAnchors calculates its average similarity score with its k.score nearest neighbors.
This score is used as a measure of the consistency and reliability
of the anchor point, with a high score
indicating that its neighbors have a high degree of
consistency across datasets.}

\item{outfeatures}{bad genes}

\item{use_old_embeddings}{}
}
\description{
Title
A Wrapper of seurat integrated, reduce and cluster
}
\examples{
\dontrun{
data("bad_genes")
seob = RunSeurat(seob = seob,
                 split_by="sample",
                 integrated="RPCA",
                 outfeatures = bad_genes,
                 coln_add = "all",
                 cellid=NULL,
                 nfeatures=3000,
                 min_cells = 3,
                 min_features = 200,
                 npcs=20,
                 resolution=0.15,
                 k_anchor=3,
                 k_weight = 200,
                 use_old_RNA = F,
                 use_old_Seurat = F)
}


}
