% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Runslingshot}
\alias{Runslingshot}
\title{Title Inferring Trajectories with Slingshot}
\usage{
Runslingshot(
  seob,
  start = NULL,
  end = NULL,
  reduction = "umap",
  cluster = "seurat_clusters"
)
}
\arguments{
\item{cluster}{}
}
\description{
Title Inferring Trajectories with Slingshot
}
