% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seurat2anndata}
\alias{seurat2anndata}
\title{Title seurat2anndata convert seurat to adata}
\usage{
seurat2anndata(
  obj,
  src = NULL,
  outFile = NULL,
  slot = "counts",
  main_layer = "RNA",
  transfer_layers = c("spliced", "unspliced", "ambiguous"),
  drop_single_values = TRUE,
  uns = NULL,
  obsm_add = NULL,
  obsm_add_name = NULL
)
}
\arguments{
\item{obsm_add_name}{}
}
\description{
Title seurat2anndata convert seurat to adata
}
\examples{
\dontrun{
bm_integrated_E11_E14 = seurat2anndata(obj = bm_integrated_E11_E14,
outFile = "data/E12_E14_WT_MT_CNCC_scVelo_scRNA_seq_E11_E14.h5ad",
slot = 'counts', main_layer = 'RNA',
transfer_layers = c("spliced", "unspliced", "ambiguous"),
drop_single_values = FALSE)
}
}
