% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{trajectory_DAG_}
\alias{trajectory_DAG_}
\title{Title a wrap function to prepare transmit net data for plot_trajectory_DAG}
\usage{
trajectory_DAG_(Tab_l, cut_, trans_link)
}
\arguments{
\item{Tab_l}{samilar to plot_trajectory_DAG}
}
\description{
Title a wrap function to prepare transmit net data for plot_trajectory_DAG
}
